Refactor resource cleaners. (#927)
Separated the cleaners into packages based on type.
